Fix a small conditional bug from last update
Since the last update, if the picture of the day is not in folder BingWallpapers in the first place, an error occurs: 
`[Errno 2] No such file or directory: '//home/USER//BingWallpapers/20161119.jpg'`
This is because of the order of the conditions introduced in the last update. I have switched it so that the conditional clause from last update (which was the first one) is now the second one. I also added some line spaces, just for readability.